Final Product of Storing Names and Sorting
This script sorts and stores names. It operates with arrays only. Call "getFinalFromStart" and pass it an array of Strings to get the final product. It will order the names by last name in ascending order but have the first name first followed by the last name. For example, "Stalin, Joseph" becomes "Joseph, Stalin."